requirements.txt
scikit-learn>=1.3.0
confluent-kafka>=2.0.0
slack-sdk>=3.21.0
python-dotenv>=1.0.0
yfinance>=0.2.29
aiohttp>=3.8.0
streamlit>=1.29.0
pandas>=2.1.0
joblib>=1.3.0
plotly>=5.19.0
ipykernel
seaborn
matplotlib